+++ animation_nodes/nodes/color/c_utils.pyx
@@ -5,12 +5,16 @@ from ... math cimport (
     hsva_to_rgba,
     hsla_to_rgba,
     yiqa_to_rgba,
+    rgba_to_hsva,
+    rgba_to_hsla,
+    rgba_to_yiqa,
 )
 from ... data_structures cimport (
     Color,
     LongList,
     FloatList,
     ColorList,
+    DoubleList,
     VirtualColorList,
     VirtualDoubleList,
     PolygonIndicesList
@@ -54,6 +58,52 @@ def colorListFromYIQA(Py_ssize_t amount, VirtualDoubleList y, VirtualDoubleList 
         yiqa_to_rgba(output.data + j, <float>y.get(j), <float>i.get(j),
                                       <float>q.get(j), <float>a.get(j))
     return output
+
+def RGBAFromColorList(ColorList colorList):
+    cdef DoubleList r = DoubleList(length = colorList.length)
+    cdef DoubleList g = DoubleList(length = colorList.length)
+    cdef DoubleList b = DoubleList(length = colorList.length)
+    cdef DoubleList a = DoubleList(length = colorList.length)
+    cdef Py_ssize_t i
+    for i in range(colorList.length):
+        r.data[i] = <double>colorList.data[i].r
+        g.data[i] = <double>colorList.data[i].g
+        b.data[i] = <double>colorList.data[i].b
+        a.data[i] = <double>colorList.data[i].a
+    return r, g, b, a
+
+def HSVAFromColorList(ColorList colorList):
+    cdef DoubleList h = DoubleList(length = colorList.length)
+    cdef DoubleList s = DoubleList(length = colorList.length)
+    cdef DoubleList v = DoubleList(length = colorList.length)
+    cdef DoubleList a = DoubleList(length = colorList.length)
+    cdef Py_ssize_t i
+    for i in range(colorList.length):
+        rgba_to_hsva(colorList.data + i, h.data + i, s.data + i,
+                     v.data + i, a.data + i)
+    return h, s, v, a
+
+def HSLAFromColorList(ColorList colorList):
+    cdef DoubleList h = DoubleList(length = colorList.length)
+    cdef DoubleList s = DoubleList(length = colorList.length)
+    cdef DoubleList l = DoubleList(length = colorList.length)
+    cdef DoubleList a = DoubleList(length = colorList.length)
+    cdef Py_ssize_t i
+    for i in range(colorList.length):
+        rgba_to_hsla(colorList.data + i, h.data + i, s.data + i,
+                     l.data + i, a.data + i)
+    return h, s, l, a
+
+def YIQAFromColorList(ColorList colorList):
+    cdef DoubleList y = DoubleList(length = colorList.length)
+    cdef DoubleList i = DoubleList(length = colorList.length)
+    cdef DoubleList q = DoubleList(length = colorList.length)
+    cdef DoubleList a = DoubleList(length = colorList.length)
+    cdef Py_ssize_t j
+    for j in range(colorList.length):
+        rgba_to_yiqa(colorList.data + j, y.data + j, i.data + j,
+                     q.data + j, a.data + j)
+    return y, i, q, a
 
 def getLoopColorsFromVertexColors(PolygonIndicesList polygons, VirtualColorList colors):
     cdef long loopsCount = polygons.indices.length