+++ CHANGELOG.md
@@ -32,6 +32,7 @@
 ### Changed
 
 - Removed Material Indices input from *Mesh Object Output* node.
+- Vectorized *Separate Color* node.
 
 # 2.1